+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.4.5] - 2024-09-18
+
+### Fixed
+
+- Fixed issue with discoverers not being found in some cases
+
 ## [1.0.3.3] - 2024-09-10
 
 ### Changed
@@ -37,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modified to use custom GA workflows
 - Published to NuGet
 
-[unreleased]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.0.3.3...HEAD
+[unreleased]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.0.4.5...HEAD
+[1.0.4.5]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.0.3.3...1.0.4.5
 [1.0.3.3]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.0.2.2...1.0.3.3
 [1.0.2.2]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.0.1.2...1.0.2.2
 [1.0.1.2]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.0.0.1...1.0.1.2